+++ CHANGELOG.md
@@ -14,6 +14,13 @@ Current Trunk
 
 ### BREAKING CHANGES (old to new)
 
+v.55
+====
+- `RelooperCreate` in the C API now has a Module parameter, and `RelooperRenderAndDispose` does not.
+
+older
+=====
+
 - `BinaryenSetFunctionTable` in the C API no longer accepts an array of functions, instead it accepts an array of function names, `const char** funcNames`. Previously, you could not include imported functions because they are of type `BinaryenImportRef` instead of `BinaryenFunctionRef`. #1650
 
 - `BinaryenSetFunctionTable` in the C API now expects the initial and maximum table size as additional parameters, like `BinaryenSetMemory` does for pages, so tables can be grown dynamically. #1687